"Added environment variables for database, storage, and message broker in README.md"
This commit adds the following environment variables to the .env file in the root directory of the project:

* RABBITMQ\_URL: the URL for the RabbitMQ message broker
* MINIO\_URL: the URL for the MinIO object storage service
* MINIO\_ACCESS\_KEY: the access key for the MinIO service
* MINIO\_SECRET\_KEY: the secret key for the MinIO service
* SECRET\_KEY: the secret key for authentication
* DATABASE\_URL: the URL for the PostgreSQL database
* MONGODB\_URL: the URL for the MongoDB database

These variables are used in the docker-compose.yml file to configure the services for the project. Adding them to the README.md file makes it easier for other developers to set up and run the project.